recover.c
We are dealing with memory management! We were given a flash drive with "deleted" files, but in reality the pointer was deleted and the images on this flash drive are still available. I created a program that would go in and recover all of the jpg taking in the file, and then outputting images of each picture stored on the drive, hence, recovering them